requirements.txt
Django==2.1
django-configurations==2.1
django-extensions==2.0.7
pytz==2018.5
six==1.11.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
djangorestframework==3.8.2
Markdown==2.6.11
Pygments==2.2.0
django-filter==1.1.0